refactor: rename float track
