+++ packages/effects-core/src/plugins/timeline/playable-assets/float-property-playable-asset.ts
@@ -2,16 +2,16 @@ import { createValueGetter } from '../../../math/value-getter';
 import { effectsClass, serialize } from '../../../decorators';
 import type { Playable, PlayableGraph } from '../../cal/playable-graph';
 import { PlayableAsset } from '../../cal/playable-graph';
-import { CurvePropertyClipPlayable } from '../playables/curve-property-clip-playable';
+import { FloatPropertyClipPlayable } from '../playables/float-property-clip-playable';
 import type { FixedNumberExpression } from '@galacean/effects-specification';
 
-@effectsClass('CurvePropertyPlayableAsset')
-export class CurvePropertyPlayableAsset extends PlayableAsset {
+@effectsClass('FloatPropertyPlayableAsset')
+export class FloatPropertyPlayableAsset extends PlayableAsset {
   @serialize()
   curveData: FixedNumberExpression;
 
   override createPlayable (graph: PlayableGraph): Playable {
-    const clipPlayable = new CurvePropertyClipPlayable(graph);
+    const clipPlayable = new FloatPropertyClipPlayable(graph);
 
     clipPlayable.curve = createValueGetter(this.curveData);
     clipPlayable.value = clipPlayable.curve.getValue(0);
